+++ metaapp/ContainerConfig.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Linq;
 using System.Text;
+using Metaapp.UI;
 
 namespace Metaapp
 {
